+++ frontend/src/pages/Main/Main.js
@@ -7,7 +7,6 @@ import ConfirmationContainer from "../Confirmation/ConfirmationContainer";
 import NotFound from "../Error/NotFound";
 import withInitialLoading from "../Loading/withInitialLoading";
 import { initLanguage } from "../Login/actions";
-import { storeSearchBarDisplayed, storeSearchTerm } from "../Navbar/actions";
 import MainNavbarNavigation from "../Navbar/MainNavbarNavigation";
 import NavbarContainer from "../Navbar/NavbarContainer";
 import NodesContainer from "../Nodes/NodesContainer";
@@ -33,14 +32,7 @@ const Main = (props) => {
       <div className="main-image" />
       <div className="main-nav">
         <NavbarContainer />
-        <MainNavbarNavigation
-          history={props.history}
-          route={props.route}
-          currentProject={props.currentProject}
-          currentSubProject={props.currentSubProject}
-          storeSearchBarDisplayed={props.storeSearchBarDisplayed}
-          storeSearchTerm={props.storeSearchTerm}
-        />
+        <MainNavbarNavigation />
       </div>
       <div className="main-container">
         <ConfirmationContainer />
@@ -68,26 +60,14 @@ class MainContainer extends Component {
   }
 
   render() {
-    return <Main {...this.props} />;
+    return <Main />;
   }
 }
 
 const mapDispatchToProps = (dispatch) => {
   return {
-    initLanguage: () => dispatch(initLanguage()),
-    storeSearchBarDisplayed: (searchBarDisplayed) => dispatch(storeSearchBarDisplayed(searchBarDisplayed)),
-    storeSearchTerm: (searchTerm) => dispatch(storeSearchTerm(searchTerm))
+    initLanguage: () => dispatch(initLanguage())
   };
 };
 
-const mapStateToProps = (state) => {
-  return {
-    currentProject: state.getIn(["navbar", "currentProject"]),
-    currentSubProject: state.getIn(["navbar", "currentSubProject"]),
-    route: state.getIn(["route", "locationBeforeTransitions"]),
-    searchBarDisplayed: state.getIn(["navbar", "searchBarDisplayed"]),
-    searchTerm: state.getIn(["navbar", "searchTerm"])
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(MainContainer);
+export default connect(null, mapDispatchToProps)(MainContainer);
