+++ frontend/src/pages/Navbar/MainNavbarNavigation.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { connect } from "react-redux";
 import { useNavigate } from "react-router-dom";
 
 import ChevronRight from "@mui/icons-material/ChevronRight";
@@ -6,6 +7,8 @@ import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 
 import strings from "../../localizeStrings";
+
+import { storeSearchBarDisplayed, storeSearchTerm } from "./actions";
 
 import "./MainNavbarNavigation.scss";
 
@@ -112,4 +115,21 @@ const MainNavbarNavigation = ({
   );
 };
 
-export default MainNavbarNavigation;
+const mapDispatchToProps = (dispatch) => {
+  return {
+    storeSearchBarDisplayed: (searchBarDisplayed) => dispatch(storeSearchBarDisplayed(searchBarDisplayed)),
+    storeSearchTerm: (searchTerm) => dispatch(storeSearchTerm(searchTerm))
+  };
+};
+
+const mapStateToProps = (state) => {
+  return {
+    currentProject: state.getIn(["navbar", "currentProject"]),
+    currentSubProject: state.getIn(["navbar", "currentSubProject"]),
+    route: state.getIn(["route", "locationBeforeTransitions"]),
+    searchBarDisplayed: state.getIn(["navbar", "searchBarDisplayed"]),
+    searchTerm: state.getIn(["navbar", "searchTerm"])
+  };
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(MainNavbarNavigation);
